requirements.txt
python-telegram-bot[webhooks]==20.7
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
httpx==0.25.2
APScheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.1
